fix: include all pops in plot
yes this is ugly. the correct way to solve this is to create a single Result file (with info that we find sensible) that specifies exactly which populations were simulated. even with that, it's not obvious to know which pops are paired (_p/_n) and which are single, so for now, this is much much simpler.