feat(multiproc): enable multithreading for NRP mode
in my testing, MUSIC is only compatible with MPI, NRP is more suited for openMP (performance improves with zero changes). so we'll keep going like this. the multiprocessing side is a bit clumped together, but... i think it's fine as is. hopefully we'll get to phase MUSIC out and further streamline.