feat(robo): draft include plane for better background
in-game camera capture does not have  any background, so the white from the skeleton is difficult to compare to the white of the background. this plane is currently placed incorrectly, but VNC isn't working right so i can't see the pybullet GUI. will take a couple commits but we'll get there!